Move response types to types/ folder
